Move ArgType and ArgNodeType to Parser
