+++ src/defineEnvironment.js
@@ -1,9 +1,9 @@
 // @flow
 import {_htmlGroupBuilders, _mathmlGroupBuilders} from "./defineFunction";
 
-import type Parser from "./Parser";
+import type Parser, {ArgType} from "./Parser";
 import type {AnyParseNode} from "./parseNode";
-import type {ArgType, Mode} from "./types";
+import type {Mode} from "./types";
 import type {NodeType} from "./parseNode";
 import type {HtmlBuilder, MathMLBuilder} from "./defineFunction";
